:hammer: code-push command becomes code-push release subcommand
